Add multiforwarder sendfd test
